Cleaned up routes: put all Get routes previously in Pages.js file into their respective [entity].js files, and renamed pages.js to index.js to hold home page router.get
