+++ src/routes/invoicedetails.js
@@ -2,6 +2,45 @@ const express = require('express');
 const router = express.Router();
 var db = require('../database/db-connector');
 
+
+// DISPLAY/READ/GET
+
+router.get('/invoicedetails', (req, res) => {
+    // join invoice details on plants by plantID and invoices by invoiceID and then gardeners by gardenerID, and then sort by invoice details ID
+    let query1 = "SELECT InvoiceDetails.invoiceDetailID, Invoices.invoiceID, Plants.varietyName, Plants.type, InvoiceDetails.quantity, InvoiceDetails.price, InvoiceDetails.lineTotal, Gardeners.firstName, Gardeners.lastName FROM InvoiceDetails INNER JOIN Plants ON InvoiceDetails.plantID = Plants.plantID INNER JOIN Invoices ON InvoiceDetails.invoiceID = Invoices.invoiceID INNER JOIN Gardeners ON Invoices.gardenerID = Gardeners.gardenerID ORDER BY InvoiceDetails.invoiceDetailID ASC;";
+    
+    // select all plants from Plants table to show in drop down
+    let query2 = "SELECT plantID, varietyName, type, price FROM Plants;";
+
+    // select all invoices from Invoices table and join with Gardeners on gardenerID to show in drop down
+    let query3 = "SELECT Invoices.invoiceID, Gardeners.gardenerID, Gardeners.firstName, Gardeners.lastName FROM Invoices INNER JOIN Gardeners ON Invoices.gardenerID = Gardeners.gardenerID ORDER BY Invoices.invoiceID ASC;";
+
+    // execute queries
+    db.pool.query(query1, function(error, rows, fields){
+        if (error) {
+            console.log(error);
+            res.sendStatus(400);
+        } else {
+            db.pool.query(query2, function(error2, plantIDs) {
+                if (error2) {
+                    console.log(error2);
+                    res.sendStatus(400);
+                } else {
+                    db.pool.query(query3, function(error3, invoiceIDs) {
+                        if (error3) {
+                            console.log(error3);
+                            res.sendStatus(400);
+                        } else {
+                            res.render('invoicedetails', { data: rows, plantIDs, invoiceIDs });
+                        }
+                    });
+                }
+            });
+        }
+    });
+});
+
+
 // CREATE/INSERT/POST
 
 router.post('/add-invoicedetail-ajax', function (req, res) {
