+++ src/routes/plantsplots.js
@@ -2,6 +2,46 @@ const express = require('express');
 const router = express.Router();
 var db = require('../database/db-connector');
 
+//DISPLAY/READ/GET
+
+router.get('/plantsplots', (req, res) => {
+    let query1 = "SELECT PlantsPlots.plantsPlotsID, Plots.plotID, Plants.plantID, Plants.varietyName, Plants.type, Gardens.gardenName FROM PlantsPlots INNER JOIN Plants ON PlantsPlots.plantID = Plants.plantID INNER JOIN Plots ON PlantsPlots.plotID = Plots.plotID INNER JOIN Gardens ON Plots.gardenID = Gardens.gardenID ORDER BY PlantsPlots.plantsPlotsID ASC;";
+
+    // select plotID, gardenName from plots joined with gardens on gardenID
+    let query2 = "SELECT Plots.plotID, Gardens.gardenName FROM Plots INNER JOIN Gardens ON Plots.gardenID = Gardens.gardenID ORDER BY Plots.plotID ASC;";
+
+    // select plants to show in dropdown
+    let query3 = "SELECT plantID, varietyName, type FROM Plants ORDER BY plantID ASC;";
+
+    db.pool.query(query1, function (error, rows, fields) {
+        if (error) {
+            console.log(error);
+            res.sendStatus(400);
+        }
+        else {
+            db.pool.query(query2, function (error2, plotIDs) {
+                if (error2) {
+                    console.log(error2);
+                    res.sendStatus(400);
+                }
+                else {
+                    db.pool.query(query3, function (error3, plantIDs) {
+                        if (error3) {
+                            console.log(error3);
+                            res.sendStatus(400);
+                        }
+                        else {
+                            res.render('plantsplots', { data: rows, plotIDs, plantIDs });
+                        }
+                    });
+                }
+            });
+        }
+
+    });
+});
+
+
 // CREATE
 router.post('/add-plantplot-ajax', function (req, res) {
     let data = req.body;
@@ -17,7 +57,7 @@ router.post('/add-plantplot-ajax', function (req, res) {
             console.log(error)
             res.sendStatus(400);
         }
-        else {            
+        else {
             db.pool.query(query2, function (error, rows, fields) {
                 if (error) {
                     console.log(error);
