+++ src/routes/plotsgardeners.js
@@ -2,6 +2,42 @@ const express = require('express');
 const router = express.Router();
 var db = require('../database/db-connector');
 
+
+// DISPLAY/READ/GET
+
+router.get('/plotsgardeners', (req, res) => {
+    // create query to join plots and gardens on gardenID, and join gardeners from PlotsGardeners on plotID
+    let query1 = "SELECT PlotsGardeners.plotsGardenersID, Plots.plotID, Gardens.gardenID, Gardeners.gardenerID, Gardens.gardenName, Gardeners.firstName, Gardeners.lastName FROM Plots INNER JOIN Gardens ON Plots.gardenID = Gardens.gardenID INNER JOIN PlotsGardeners ON Plots.plotID = PlotsGardeners.plotID INNER JOIN Gardeners ON PlotsGardeners.gardenerID = Gardeners.gardenerID ORDER BY PlotsGardeners.plotsGardenersID ASC;";
+    // query all plots and garden name from joined plots and gardens
+    let query2 = "SELECT Plots.plotID, Gardens.gardenName FROM Plots INNER JOIN Gardens ON Plots.gardenID = Gardens.gardenID ORDER BY Plots.plotID ASC;";
+    // query all gardeners
+    let query3 = "SELECT gardenerID, firstName, lastName FROM Gardeners;";
+
+    db.pool.query(query1, function (error, rows, fields) {
+        if (error) {
+            console.log(error);
+            res.sendStatus(400);
+        } else {
+            db.pool.query(query2, function (error2, plotIDs) {
+                if (error2) {
+                    console.log(error2);
+                    res.sendStatus(400);
+                } else {
+                    db.pool.query(query3, function (error3, gardenerIDs) {
+                        if (error3) {
+                            console.log(error3);
+                            res.sendStatus(400);
+                        } else {
+                            res.render('plotsgardeners', { data: rows, plotIDs, gardenerIDs });
+                        }
+                    });
+                }
+            });
+        }
+    });
+});
+
+
 // CREATE
 router.post('/add-plotgardener-ajax', function (req, res) {
     let data = req.body;
@@ -9,7 +45,7 @@ router.post('/add-plotgardener-ajax', function (req, res) {
 
     // insert
     let query1 = `INSERT INTO PlotsGardeners (plotID, gardenerID) VALUES ('${data.plotID}', '${data.gardenerID}');`;
-    
+
     // select to update the table
     let query2 = "SELECT PlotsGardeners.plotsGardenersID, Plots.plotID, Gardens.gardenID, Gardeners.gardenerID, Gardens.gardenName, Gardeners.firstName, Gardeners.lastName FROM Plots INNER JOIN Gardens ON Plots.gardenID = Gardens.gardenID INNER JOIN PlotsGardeners ON Plots.plotID = PlotsGardeners.plotID INNER JOIN Gardeners ON PlotsGardeners.gardenerID = Gardeners.gardenerID ORDER BY PlotsGardeners.plotsGardenersID ASC;";
 
@@ -18,7 +54,7 @@ router.post('/add-plotgardener-ajax', function (req, res) {
             console.log(error)
             res.sendStatus(400);
         }
-        else {            
+        else {
             db.pool.query(query2, function (error, rows, fields) {
                 if (error) {
                     console.log(error);
